not_railway/requirements-slim.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==10.0.0
pdf2image==1.16.3
numpy==1.24.3
scikit-learn==1.3.0